feature: adding the recovery management undo-redo algorithm

+++ src/main/java/com/luka/simpledb/transactionManagement/recoveryManagement/RecoveryManager.java
@@ -18,6 +18,7 @@ public class RecoveryManager {
     private final BufferManager bufferManager;
     private final Transaction transaction;
     private final int transactionNumber;
+    private final RecoveryAlgorithm recoveryAlgorithm = new RecoveryAlgorithm(true);
 
     /// A recovery manager is instantiated with a given transaction, and is
     /// tied to it until the end of its lifecycle. A buffer manager is needed
@@ -35,14 +36,15 @@ public class RecoveryManager {
         StartRecord.writeToLog(logManager, transactionNumber);
     }
 
-    /// Commits the tied transaction. Since the recovery algorithm implemented by the
-    /// recovery step is the 'Undo-Only' algorithm, the commit log record must be written
-    /// **after** all buffers used by the transaction are flushed, because that is what
-    /// the recovery algorithm assumes (the assumption is that all commited transactions
-    /// will have their changes written out in user data blocks meaning they won't need
-    /// to be reapplied on system recovery). This has a penalty on performance.
+    /// Commits the tied transaction.
     public void commit() {
-        bufferManager.flushAll(transactionNumber);
+        if (recoveryAlgorithm.undoOnly) {
+            // the commit log record must be written **after** all buffers used by the transaction
+            // are flushed, because that is what the 'Undo-Only' algorithm assumes (the assumption is
+            // that all commited transactions will have their changes written out in user data blocks
+            // meaning they won't need to be reapplied on system recovery); this has a penalty on performance
+            bufferManager.flushAll(transactionNumber);
+        }
         int lsn = CommitRecord.writeToLog(logManager, transactionNumber);
         logManager.flush(lsn);
     }
@@ -65,34 +67,30 @@ public class RecoveryManager {
     /// performed are written to the disk and a quiescent checkpoint can be written out since
     /// the system is sure that the database is in a reasonable state.
     public void recover() {
-        undoOnlyRecover();
+        recoveryAlgorithm.systemRecovery();
         bufferManager.flushAll(transactionNumber);
         int lsn = QuiescentCheckpointRecord.writeToLog(logManager);
         logManager.flush(lsn);
     }
 
     /// Creates an update log for updating an integer in a given buffer on a given offset.
-    /// The parameter `newValue` is not used when the system is configured to 'Undo-Only'
-    /// recovery to save space.
     ///
     /// @return The log sequence number for the record in the log file corresponding to the
     /// call of this function.
     public int setInt(Buffer buffer, int offset, int newValue) {
         int oldValue = buffer.getContents().getInt(offset);
         BlockId blockId = buffer.getBlockId();
-        return SetIntRecord.writeToLog(logManager, transactionNumber, blockId, offset, oldValue);
+        return SetIntRecord.writeToLog(logManager, transactionNumber, blockId, offset, oldValue, newValue);
     }
 
     /// Creates an update log for updating a string in a given buffer on a given offset.
-    /// The parameter `newValue` is not used when the system is configured to 'Undo-Only'
-    /// recovery to save space.
     ///
     /// @return The log sequence number for the record in the log file corresponding to the
     /// call of this function.
     public int setString(Buffer buffer, int offset, String newValue) {
         String oldValue = buffer.getContents().getString(offset);
         BlockId blockId = buffer.getBlockId();
-        return SetStringRecord.writeToLog(logManager, transactionNumber, blockId, offset, oldValue);
+        return SetStringRecord.writeToLog(logManager, transactionNumber, blockId, offset, oldValue, newValue);
     }
 
     /// The algorithm for rolling back a transaction. The transaction that
@@ -116,41 +114,88 @@ public class RecoveryManager {
         }
     }
 
-    /// The 'Undo-Only' algorithm for system recovery.
-    private void undoOnlyRecover() {
-        // remember all completed transactions
-        HashSet<Integer> finishedTransactions = new HashSet<>();
-        Iterator<byte[]> iter = logManager.iterator();
+    /// Types of algorithms used for system recovery.
+    class RecoveryAlgorithm {
+        private final ArrayList<LogRecord> processedRecords = new ArrayList<>();
+        private final HashSet<Integer> completedTransactions = new HashSet<>();
+        public final boolean undoOnly;
 
-        while (iter.hasNext()) {
-            byte[] bytes = iter.next();
-            LogRecord record = LogRecord.createLogRecord(bytes);
-            if (record.op() == LogRecordType.QUIESCENT_CHECKPOINT) {
-                // if a quiescent checkpoint is found, that means the
-                // log doesn't need to be traversed any further because
-                // the database is guaranteed to be in a reasonable state
-                // in correspondence to the logs before the quiescent checkpoint
-                // log
-                return;
-            }
-            if (record.op() == LogRecordType.COMMIT || record.op() == LogRecordType.ROLLBACK) {
-                // a transaction is completed if it's rollback or commit log
-                // is found in the log file
-                finishedTransactions.add(record.transactionNumber());
-            } else if (!finishedTransactions.contains(record.transactionNumber())) {
-                // if some other than commit or rollback log record type
-                // is encountered, and its corresponding transaction is not
-                // completed (the traversing is done from the end of the file,
-                // so commit and rollback logs will always be encountered first)
-                // it means that record must be undone since a reasonable state is
-                // one where all uncompleted transactions are rolled back
-                record.undo(transaction);
+        /// When `undoOnly` is set to `false`, the algorithm
+        /// used is the 'Undo-Redo' algorithm.
+        public RecoveryAlgorithm(boolean undoOnly) {
+            this.undoOnly = undoOnly;
+        }
+
+        /// Recover the system if it crashed. The recovery is an idempotent
+        /// function so if the system crashes in the middle of recovery,
+        /// the system won't be left in an inconsistent state.
+        public void systemRecovery() {
+            if (undoOnly) {
+                undoOnlyRecover();
+            } else {
+                undoRedoRecover();
             }
         }
-    }
 
-    /// The 'Undo-Redo' algorithm for system recovery.
-    private void undoRedoRecover() {
-        // todo
+        /// The 'Undo-Only' algorithm for system recovery. Assumes all commited transactions
+        /// will have their changes written out in user data blocks meaning they won't need
+        /// to be reapplied on system recovery.
+        /// Advantage: faster recovery. Disadvantage: slower commits.
+        private void undoOnlyRecover() {
+            Iterator<byte[]> iter = logManager.iterator();
+
+            while (iter.hasNext()) {
+                byte[] bytes = iter.next();
+                LogRecord record = LogRecord.createLogRecord(bytes);
+                if (record.op() == LogRecordType.QUIESCENT_CHECKPOINT) {
+                    // if a quiescent checkpoint is found, that means the
+                    // log doesn't need to be traversed any further because
+                    // the database is guaranteed to be in a reasonable state
+                    // in correspondence to the logs before the quiescent checkpoint
+                    // log
+                    return;
+                }
+
+                if (!undoOnly) {
+                    // do not use unnecessary space when doing only
+                    // the undo part of the recovery
+                    processedRecords.add(record);
+                }
+
+                if (record.op() == LogRecordType.COMMIT || record.op() == LogRecordType.ROLLBACK) {
+                    // a transaction is completed if it's rollback or commit log
+                    // is found in the log file
+                    completedTransactions.add(record.transactionNumber());
+                } else if (!completedTransactions.contains(record.transactionNumber())) {
+                    // if some other than commit or rollback log record type
+                    // is encountered, and its corresponding transaction is not
+                    // completed (the traversing is done from the end of the file,
+                    // so commit and rollback logs will always be encountered first)
+                    // it means that record must be undone since a reasonable state is
+                    // one where all uncompleted transactions are rolled back
+                    record.undo(transaction);
+                }
+            }
+        }
+
+        /// The 'Undo-Redo' algorithm for system recovery.
+        /// Disadvantage: slow recovery, uses more memory.
+        private void undoRedoRecover() {
+            // firstly, call the undo portion of the algorithm
+            undoOnlyRecover();
+
+            // redo every record, but in reverse order
+            for (LogRecord record : processedRecords.reversed()) {
+                if (!completedTransactions.contains(record.transactionNumber())) {
+                    // if some other than commit or rollback log record type
+                    // is encountered, and its corresponding transaction is
+                    // completed (list of completed transactions already exists from
+                    // the backward pass in the undo portion) it means that record
+                    // must be redone since a reasonable state is one where all
+                    // completed transactions are commited
+                    record.redo(transaction);
+                }
+            }
+        }
     }
 }
